requirements.txt
tensorflow
pandas
numpy
matplotlib
scikit-learn
streamlit
Pillow
